+++ packages/meteor-tool/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "The Meteor command-line tool",
-  version: '1.4.2_1'
+  version: '1.4.2_3'
 });
 
 Package.includeTool();