+++ packages/twitter/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Twitter OAuth flow",
-  version: '1.1.13'
+  version: '1.1.14'
 });
 
 Package.onUse(function(api) {