+++ tools/cli/commands.js
@@ -770,6 +770,11 @@ main.registerCommand({
   }
 
   Console.info(
+    Console.command("meteor npm install"),
+    Console.options({ indent: 2 }),
+  );
+
+  Console.info(
     Console.command("meteor"), Console.options({ indent: 2 }));
 
   Console.info("");